(ios) Removing "max fee" configuration option. Moving "background payments" configuration from "display options" to "payment options"
