+++ phoenix-ios/phoenix-ios/views/configuration/ConfigurationView.swift
@@ -442,7 +442,7 @@ fileprivate struct ConfigurationList: View {
 				case .backup             : newNavLinkTag = .RecoveryPhrase       ; delay *= 1
 				case .drainWallet        : newNavLinkTag = .DrainWallet          ; delay *= 1
 				case .electrum           : newNavLinkTag = .ElectrumServer       ; delay *= 1
-				case .backgroundPayments : newNavLinkTag = .DisplayConfiguration ; delay *= 2
+				case .backgroundPayments : newNavLinkTag = .PaymentOptions       ; delay *= 2
 				case .liquiditySettings  : newNavLinkTag = .ChannelManagement    ; delay *= 1
 				case .forceCloseChannels : newNavLinkTag = .ForceCloseChannels   ; delay *= 1
 			}
